@@ -1,37 +1,21 @@
 package work.lclpnet.notica.network.packet;
 
-import net.fabricmc.fabric.api.networking.v1.FabricPacket;
-import net.fabricmc.fabric.api.networking.v1.PacketType;
 import net.minecraft.network.PacketByteBuf;
+import net.minecraft.network.codec.PacketCodec;
+import net.minecraft.network.packet.CustomPayload;
 import net.minecraft.util.Identifier;
 import work.lclpnet.notica.NoticaInit;
 
-public class StopSongBidiPacket implements FabricPacket {
+public record StopSongBidiPacket(Identifier songId) implements CustomPayload {
 
-    public static final PacketType<StopSongBidiPacket> TYPE =
-            PacketType.create(NoticaInit.identifier("stop"), StopSongBidiPacket::new);
+    public static final Id<StopSongBidiPacket> ID = new Id<>(NoticaInit.identifier("stop"));
 
-    private final Identifier songId;
-
-    public StopSongBidiPacket(Identifier songId) {
-        this.songId = songId;
-    }
-
-    public StopSongBidiPacket(PacketByteBuf buf) {
-        this.songId = buf.readIdentifier();
-    }
+    public static final PacketCodec<PacketByteBuf, StopSongBidiPacket> CODEC = PacketCodec.tuple(
+            Identifier.PACKET_CODEC, StopSongBidiPacket::songId,
+            StopSongBidiPacket::new);
 
     @Override
-    public void write(PacketByteBuf buf) {
-        buf.writeIdentifier(songId);
-    }
-
-    @Override
-    public PacketType<?> getType() {
-        return TYPE;
-    }
-
-    public Identifier getSongId() {
-        return songId;
+    public Id<? extends CustomPayload> getId() {
+        return ID;
     }
 }
